Create parsers test file
